+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+Development
+-----------
+
+- [Update] Move all in-memory state to the DB. And allow listing of previous upload information.
+  (`MR124 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/124>`__)
+
 0.4.0
 -----